refactor(source): split interface file into smaller perimeter - SRC-4641
