+++ src/resources/Sources/SourcesSubInterfaces/Address.ts
@@ -1,0 +1,7 @@
+import {PatternType} from '../../Enums';
+
+export interface AddressPattern {
+    allowed?: boolean;
+    expression?: string;
+    patternType?: PatternType;
+}
