+++ src/resources/Sources/SourcesSubInterfaces/Error.ts
@@ -1,0 +1,4 @@
+export interface ConfigurationError {
+    errorCode?: string;
+    message?: string;
+}
