+++ src/resources/Sources/SourcesSubInterfaces/ExtendedDataFile.ts
@@ -1,0 +1,5 @@
+export interface ExtendedDataFile {
+    data?: string;
+    extendedData?: any;
+    sensitive?: boolean;
+}
