+++ src/resources/Sources/SourcesSubInterfaces/Extension.ts
@@ -1,0 +1,10 @@
+import {SourceExtensionActionOnError} from '../../Enums';
+
+export interface SourceExtensionModel {
+    actionOnError?: SourceExtensionActionOnError;
+    condition?: string;
+    extensionId?: string;
+    itemType?: string;
+    parameters?: Record<string, string>;
+    versionId?: string;
+}
