+++ src/resources/Sources/SourcesSubInterfaces/Parameter.ts
@@ -1,0 +1,4 @@
+export interface ParameterModel {
+    sensitive?: boolean;
+    value?: string;
+}
