+++ src/resources/Sources/SourcesSubInterfaces/Url.ts
@@ -1,0 +1,7 @@
+import {UrlFilterType} from '../../Enums';
+
+export interface UrlFilter {
+    filter?: string;
+    filterType?: UrlFilterType;
+    includeFilter?: boolean;
+}
