+++ src/resources/Sources/SourcesSubInterfaces/User.ts
@@ -1,0 +1,5 @@
+export interface UserIdentityModel {
+    name?: string;
+    password?: string;
+    userName?: string;
+}
