+++ src/resources/Sources/SourcesSubInterfaces/index.ts
@@ -1,0 +1,11 @@
+export * from './Address';
+export * from './Authentification';
+export * from './Document';
+export * from './Error';
+export * from './ExtendedDataFile';
+export * from './Extension';
+export * from './Information';
+export * from './Parameter';
+export * from './Permission';
+export * from './Url';
+export * from './User';
